Implement test for toboggan slide

+++ app/src/main/kotlin/adventofcode2020/Day1.kt
@@ -9,11 +9,11 @@ class Day1(resource: Resource) : ResourceSolver(resource) {
        resource.lines.map { it.toInt() }.sorted()
     }
 
-    override fun solve1(): Int {
+    override fun solve1(): Long {
         TODO("Implement")
     }
 
-    override fun solve2(): Int {
+    override fun solve2(): Long {
         // solve with a fast runner
         // but let's not make it fancy yet
         var i = 0
@@ -21,7 +21,7 @@ class Day1(resource: Resource) : ResourceSolver(resource) {
             for (j in i..(list.size - 1)) {
                 for (k in (j..list.size - 1)) {
                     if (list[i] + list[j] + list[k] == 2020) {
-                        return list[i] * list[j] * list[k]
+                        return (list[i] * list[j] * list[k]).toLong()
                     }
                 }
             }
